Made Basic Player Controller
The player can now move up, down, left, right, and forward. A lot of adjustment is still needed on each value. It will probably also be a good idea to change the transform.position += 1 for the jump and just set it to a fixed height and scale the block smaller in all axes.